Move LedgerKind to comit_api
